pyproject.toml
[project]
name = "language-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "better-exchook>=1.20260303.161200",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
